Extracted mock tools out into a separate module.
